layers: Move CoreChecks to Ray Tracing file
